chore: final touches
This commit cleans up the ODK Central work, having moved some items out
of the odk-central file to be put in the ODK Central API library.
