+++ client/src/modules/odk-settings/odk-settings.html
@@ -89,7 +89,7 @@
 
             <div class="form-group">
               <label class="control-label" translate>ODK.SYNC_USERS</label>
-              <button type="button" class="btn btn-default btn-xs pull-right" ng-click="ODKSettingsCtrl.syncUsers()" translate>FORM.BUTTONS.SYNC</button>
+              <button type="button" class="btn btn-default btn-xs pull-right" ng-click="ODKSettingsCtrl.syncAppUsers()" translate>FORM.BUTTONS.SYNC</button>
             </div>
