+++ server/models/schema.sql
@@ -2644,21 +2644,9 @@ CREATE TABLE `odk_central_integration` (
   CONSTRAINT `odk_central__enterprise` FOREIGN KEY (`enterprise_id`) REFERENCES `enterprise` (`id`)
 ) ENGINE=InnoDB DEFAULT CHARACTER SET = utf8mb4 DEFAULT COLLATE = utf8mb4_unicode_ci;
 
-
--- @jniles
-DROP TABLE IF EXISTS `odk_user`;
-CREATE TABLE `odk_user` (
-  `odk_user_id` INT UNSIGNED NOT NULL,
-  `odk_user_password` TEXT NOT NULL,
-  `bhima_user_id` SMALLINT(5) UNSIGNED NOT NULL,
-  CONSTRAINT `odk_user__user` FOREIGN KEY (`bhima_user_id`) REFERENCES `user` (`id`)
-) ENGINE=InnoDB DEFAULT CHARACTER SET = utf8mb4 DEFAULT COLLATE = utf8mb4_unicode_ci;
-
--- @mbayopanda
 DROP TABLE IF EXISTS `odk_app_user`;
 CREATE TABLE `odk_app_user` (
   `odk_app_user_id` INT UNSIGNED NOT NULL,
-  `odk_app_user_token` TEXT NOT NULL,
   `display_name` TEXT NOT NULL,
   `bhima_user_id` SMALLINT(5) UNSIGNED NOT NULL,
   CONSTRAINT `odk_app_user__user` FOREIGN KEY (`bhima_user_id`) REFERENCES `user` (`id`)
